Updated SQL Database
Added Playlist Table
Due to some limitations, I thought that the best way was to have a single table with the user id and the playlist id, then have a separate table that actually held the contents of each playlist. If you guys have a better way, go for it.

PS: Idk how to do the auto incrementing stuff so if one of yall can add that to the Playlist ID under the Playlist table, that'd be great